requirements.txt
asgiref==3.8.1
Django==5.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
Faker==28.0.0
ksuid==1.3
numpy==2.1.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
